+++ storage_media/models/__init__.py
@@ -2,4 +2,4 @@
 
 from . import storage_file
 from . import storage_media
-
+from . import storage_media_type
